test: extract `tests` directory out of `src`
