+++ tests/auto-cleanup.test.js
@@ -1,6 +1,6 @@
 import { afterEach, beforeEach, describe, expect, test, vi } from 'vitest'
 
-import { IS_JEST } from './utils.js'
+import { IS_JEST } from './_env.js'
 
 // TODO(mcous, 2024-12-08): clearing module cache and re-importing
 // in Jest breaks Svelte's environment checking heuristics.
@@ -19,7 +19,7 @@ describe.skipIf(IS_JEST)('auto-cleanup', () => {
   })
 
   test('calls afterEach with cleanup if globally defined', async () => {
-    const { render } = await import('../index.js')
+    const { render } = await import('@testing-library/svelte')
 
     expect(globalAfterEach).toHaveBeenCalledTimes(1)
     expect(globalAfterEach).toHaveBeenLastCalledWith(expect.any(Function))
@@ -35,7 +35,7 @@ describe.skipIf(IS_JEST)('auto-cleanup', () => {
   test('does not call afterEach if process STL_SKIP_AUTO_CLEANUP is set', async () => {
     process.env.STL_SKIP_AUTO_CLEANUP = 'true'
 
-    await import('../index.js')
+    await import('@testing-library/svelte')
 
     expect(globalAfterEach).toHaveBeenCalledTimes(0)
   })
