+++ tests/fixtures/Comp.svelte
