+++ tests/fixtures/CompRunes.svelte
