+++ tests/fixtures/Typed.svelte
