+++ tests/fixtures/TypedRunes.svelte
