+++ tests/render-utilities.test-d.ts
@@ -1,7 +1,7 @@
+import * as subject from '@testing-library/svelte'
 import { expectTypeOf } from 'expect-type'
 import { describe, test } from 'vitest'
 
-import * as subject from '../index.js'
 import Component from './fixtures/Comp.svelte'
 
 describe('render query and utility types', () => {
