+++ tests/render.test-d.ts
@@ -1,8 +1,8 @@
+import * as subject from '@testing-library/svelte'
 import { expectTypeOf } from 'expect-type'
 import { ComponentProps } from 'svelte'
 import { describe, test } from 'vitest'
 
-import * as subject from '../index.js'
 import Component from './fixtures/Typed.svelte'
 
 describe('types', () => {
