+++ tests/transition.test.js
@@ -2,8 +2,8 @@ import { render, screen, waitFor } from '@testing-library/svelte'
 import { userEvent } from '@testing-library/user-event'
 import { beforeEach, describe, expect, test, vi } from 'vitest'
 
+import { IS_JSDOM, IS_SVELTE_5 } from './_env.js'
 import Transitioner from './fixtures/Transitioner.svelte'
-import { IS_JSDOM, IS_SVELTE_5 } from './utils.js'
 
 describe.skipIf(IS_SVELTE_5)('transitions', () => {
   beforeEach(() => {
